Extract the card graph wrapper in Disks to a separate component
