+++ packages/manager/src/features/Longview/LongviewDetail/DetailTabs/Disks/Graphs.tsx
@@ -1,34 +1,13 @@
 import { pathOr } from 'ramda';
 import * as React from 'react';
 import { compose } from 'recompose';
-import {
-  makeStyles,
-  Theme,
-  withTheme,
-  WithTheme
-} from 'src/components/core/styles';
+import { withTheme, WithTheme } from 'src/components/core/styles';
 
 import Typography from 'src/components/core/Typography';
 
 import LongviewLineGraph from 'src/components/LongviewLineGraph';
 import { StatWithDummyPoint } from '../../../request.types';
-
-const useStyles = makeStyles((theme: Theme) => ({
-  graphContainer: {
-    marginTop: theme.spacing(),
-    display: 'flex',
-    flexFlow: 'row wrap',
-    justifyContent: 'space-around',
-    '& > div': {
-      flexGrow: 1,
-      width: '33%',
-      [theme.breakpoints.down('md')]: {
-        marginTop: theme.spacing(),
-        width: '60%'
-      }
-    }
-  }
-}));
+import GraphCard from '../../GraphCard';
 
 export interface Props {
   isSwap: boolean;
@@ -50,8 +29,6 @@ export interface Props {
 type CombinedProps = Props & WithTheme;
 
 const Graphs: React.FC<CombinedProps> = props => {
-  const classes = useStyles();
-
   const {
     isSwap,
     childOf,
@@ -87,87 +64,73 @@ const Graphs: React.FC<CombinedProps> = props => {
   const labelHelperText = generateHelperText(sysInfoType, isSwap, isMounted);
 
   return (
-    <React.Fragment>
-      <Typography variant="subtitle1">
-        <React.Fragment>
-          <strong>{diskLabel}</strong>
-          {!!labelHelperText && <React.Fragment> &ndash; </React.Fragment>}
-          {labelHelperText}
-        </React.Fragment>
-      </Typography>
-      <div className={classes.graphContainer}>
-        {/* 
-            openVZ disks don't have I/O stats. This logic comes straight from 
-            old Longview.JS, so a big @todo here is to document why this
-            is the way it is.
-          */}
-        {sysInfoType.toLowerCase() !== 'openvz' && (
-          <div data-testid="diskio-graph">
-            <LongviewLineGraph
-              data={[
-                {
-                  data: formatDiskIO(writes),
-                  label: 'Write',
-                  borderColor: theme.graphs.orangeBorder,
-                  backgroundColor: theme.graphs.orange
-                },
-                {
-                  data: formatDiskIO(reads),
-                  label: 'Read',
-                  borderColor: theme.graphs.yellowBorder,
-                  backgroundColor: theme.graphs.yellow
-                }
-              ]}
-              title="Disk I/O"
-              showToday={isToday}
-              subtitle="ops/s"
-              timezone={timezone}
-            />
-          </div>
-        )}
-        {/*
+    <GraphCard title={diskLabel} helperText={labelHelperText}>
+      {sysInfoType.toLowerCase() !== 'openvz' && (
+        <div data-testid="diskio-graph">
+          <LongviewLineGraph
+            data={[
+              {
+                data: formatDiskIO(writes),
+                label: 'Write',
+                borderColor: theme.graphs.orangeBorder,
+                backgroundColor: theme.graphs.orange
+              },
+              {
+                data: formatDiskIO(reads),
+                label: 'Read',
+                borderColor: theme.graphs.yellowBorder,
+                backgroundColor: theme.graphs.yellow
+              }
+            ]}
+            title="Disk I/O"
+            showToday={isToday}
+            subtitle="ops/s"
+            timezone={timezone}
+          />
+        </div>
+      )}
+      {/*
             only show inodes and space if the
             disk is mounted and is not a swap partition
             because longview doesn't track those stats
           */
-        !isSwap && isMounted && (
-          <React.Fragment>
-            <div data-testid="space-graph">
-              <LongviewLineGraph
-                data={[
-                  {
-                    data: formatSpace(free, total),
-                    label: 'Space',
-                    borderColor: theme.graphs.salmonBorder,
-                    backgroundColor: theme.graphs.salmon
-                  }
-                ]}
-                showToday={isToday}
-                title="Space"
-                subtitle="GB"
-                timezone={timezone}
-              />
-            </div>
-            <div data-testid="inodes-graph">
-              <LongviewLineGraph
-                data={[
-                  {
-                    data: formatINodes(iFree, iTotal),
-                    label: 'Inodes',
-                    borderColor: theme.graphs.pinkBorder,
-                    backgroundColor: theme.graphs.pink
-                  }
-                ]}
-                showToday={isToday}
-                suggestedMax={1000000}
-                title="Inodes"
-                timezone={timezone}
-              />
-            </div>
-          </React.Fragment>
-        )}
-      </div>
-    </React.Fragment>
+      !isSwap && isMounted && (
+        <React.Fragment>
+          <div data-testid="space-graph">
+            <LongviewLineGraph
+              data={[
+                {
+                  data: formatSpace(free, total),
+                  label: 'Space',
+                  borderColor: theme.graphs.salmonBorder,
+                  backgroundColor: theme.graphs.salmon
+                }
+              ]}
+              showToday={isToday}
+              title="Space"
+              subtitle="GB"
+              timezone={timezone}
+            />
+          </div>
+          <div data-testid="inodes-graph">
+            <LongviewLineGraph
+              data={[
+                {
+                  data: formatINodes(iFree, iTotal),
+                  label: 'Inodes',
+                  borderColor: theme.graphs.pinkBorder,
+                  backgroundColor: theme.graphs.pink
+                }
+              ]}
+              showToday={isToday}
+              suggestedMax={1000000}
+              title="Inodes"
+              timezone={timezone}
+            />
+          </div>
+        </React.Fragment>
+      )}
+    </GraphCard>
   );
 };
 
@@ -251,7 +214,4 @@ export const generateHelperText = (
   return '';
 };
 
-export default compose<CombinedProps, Props>(
-  React.memo,
-  withTheme
-)(Graphs);
+export default compose<CombinedProps, Props>(React.memo, withTheme)(Graphs);
